+++ examples/angular/src/app/app.routes.ts
@@ -2,6 +2,7 @@ import { Routes } from '@angular/router';
 import { FormComponent } from './pages/form/form.component';
 import { HomeComponent } from './pages/home/home.component';
 import { AlertComponent } from './pages/alert/alert.component';
+import { DialogComponent } from './pages/dialog/dialog.component';
 
 export const routes: Routes = [
   {
@@ -20,5 +21,9 @@ export const routes: Routes = [
   {
     path: 'alert',
     component: AlertComponent,
+  },
+  {
+    path: 'dialog',
+    component: DialogComponent,
   },
 ];